requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
redis==5.0.1
sqlalchemy==2.0.23
prometheus-client==0.19.0
structlog==23.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
